requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==4.0.1
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
cachetools==7.0.5
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
cloudpathlib==0.23.0
colorama==0.4.6
confection==1.3.3
cymem==2.0.13
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.15.0
filelock==3.25.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.4.0
h11==0.16.0
hf-xet==1.4.3
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.10.2
idna==3.11
iniconfig==2.3.0
installer==0.7.0
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.11
langchain-pinecone==0.2.3
langchain-tests==0.3.17
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.15
langsmith==0.7.30
llvmlite==0.47.0
lxml==6.0.4
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.19.0
networkx==3.6.1
numba==0.65.0
numpy==1.26.4
olefile==0.47
openai==1.109.1
orjson==3.11.8
ormsgpack==1.12.2
packaging==24.2
pandas==3.0.2
passlib==1.7.4
pillow==12.2.0
pinecone==5.4.2
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
plotly==6.7.0
pluggy==1.6.0
preshed==3.0.13
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pyarrow==23.0.1
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
pydeck==0.9.1
Pygments==2.20.0
pypdf==6.10.1
pypdfium2==5.7.0
pytest==8.4.2
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.3
RapidFuzz==3.14.5
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
smart_open==7.6.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.49
srsly==2.5.3
streamlit==1.56.0
sympy==1.14.0
syrupy==4.9.1
tenacity==9.1.4
thinc==8.3.13
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0+cpu
torchvision==0.26.0+cpu
tornado==6.5.5
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
unstructured==0.22.20
unstructured-client==0.43.2
urllib3==2.6.3
uuid_utils==0.14.1
wasabi==1.1.3
watchdog==6.0.0
weasel==1.0.0
webencodings==0.5.1
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
